Generate a table of contents
This adds the table of contents to pages. It is dynamically created from
the h2s on the page. As the user scrolls it becomes sticky and shows the
h2 that the user is currently viewing. If the user hovers over the
header it will show the full TOC menu.