Only show the survey-ad after the user has visited a few pages
After the user visits CanJS.com for the first time, we count any additional page loads (across sessions) as “engagements.” After three engagements, we show them the survey-ad.

This shifts the control to showing unless dismissed to only showing if they have JS enabled and have interacted with the site a few times.