Replace current sidebar with new sidebar
This is a big work in progress hack; right now, the old sidebar is being replaced, but it’s done from the search control when it should be done from the main canjs.js file.

Additionally, clicking on links in the sidebar from the homepage works, but reloading and using the sidebar doesn’t work (need to fix the link URLs).

Overall, it’s pretty broken, but a step in the right direction.