Add buttons for expanding/collapsing the sections under API Docs
Each item in the sidebar now has a `isCollapsible` property that controls whether the expand/collapse button should be shown for that item. The property is only true for the pages directly under the API Docs page.

The `isCollapsed` property keeps track of whether an item is collapsed; by default, everything is collapsed; we’ll get to that in a minute. The `collapse` method flips `isCollapsed` to true or false.

In addition to `unsortedChildren` and `sortedChildren`, the new `visibleChildren` property returns the list of children that should be shown, depending on whether the page is currently collapsed or not. If `isCollapsible` and `isCollapsed` are both true, then only Core collection modules are returned.

By default, every page `isCollapsed`, but this is relevant to pages with `isCollapsible` because otherwise, `visibleChildren` will return all children.