Don’t animate the sections being expanded when it’s not user-initiated
The sections in the new sidebar are expanded via two methods: when the user explicitly initiates the section being opened (via the + button) or when a section is opened because the user clicked a link that causes the section to be opened.

In the latter situation, it’s unnecessary for the expansion animation to happen, so by default all transitions are off and we only enable transitions when the user clicks the +/i buttons.