“Fix” the GitHub star buttons on the API Docs page in IE 11
It takes the buttons script an incredibly long time to execute on IE 11 (upwards of 30 seconds), so I’m “progressively degrading” the experience for IE 11 and not upgrading the star links to buttons.

Additionally, the npm install command and buttons weren’t being shown because `display: unset` is not supported in IE, so that’s fixed.