Fix an issue with groups not being expanded when a page in a non-Core package is selected
Previously, when a page within a package that is not part of the Core collection was selected, the group that it’s in would not automatically be expanded. This bug was apparent on this page: https://canjs.com/doc/can-view-nodelist/types/NodeList.html

This commit also adds some docs about the sidebar.