Fix navigating to and from URLs with hashes
Previously, if you clicked on a link to a hash section and then clicked another link to a hash section, the second click wouldn’t work.

This is fixed by correctly determining the hash you’re navigating to and from. Additionally, this fixes an issue in Firefox where the “On this page” navigation wouldn’t be updated with the correct title, and an issue in Chrome which would cause the navigation to be hidden when using the browser’s back button.

Fixes https://github.com/canjs/bit-docs-html-canjs/issues/420