Fix issue with clicking docs in the sidebar
When the user clicks on the docs for a package within the Core collection, the purpose group that it’s in (e.g. Observables) shouldn’t expand automatically.

This fixes an issue where sections were expanding automatically when they shouldn’t, which fixed the issue with docs pages not loading correctly.

Fixes https://github.com/canjs/canjs/issues/4075